Sh_Func_L2norm translation advances

+++ Tutorials/OpenArmsTutorial/provaCall1.m
@@ -1,5 +1,6 @@
 params.designVariable = struct('thetavec', [1.0, 2.0, 3.0]);
-params.x = [1.0, 2.0, 3.0];
+params.x = double([1.0, 2.0, 3.0]);
+
 
 output = callJuliaClass('Sh_Func_L2norm', 'computeFunctionAndGradient', params);
 disp(output.j)   % cost
